requirements.txt
google-generativeai
PyPDF2
gradio